pyproject.toml
[tool.poetry]
name = "march-tales"
version = "0.0.39"
description = "March Tales django rest api and web frontend server"
authors = ["lilliputten <lilliputten@gmail.com>"]
readme = "README.md"

# @see https://python-poetry.org/docs/pyproject/#packages
packages = [
    { include = "tales_django" },
]

[tool.poetry.dependencies]
concurrent-log-handler = "^0.9.25" # Write logs to a retating files
beautifulsoup4 = "^4.12.3"
colorama = "^0.4.6"
crispy-bootstrap5 = "^2024.10"
django = "5.0"
django-compressor = "^4.5.1"
django-crispy-forms = "^2.3"
django-environ = "^0.11.2"
django-livereload-server = "^0.5.1"
django-preferences = "^1.0.0"
django-registration = "^3.4" # Using an old version intentionally # = "^5.1.0" # @see https://django-registration.readthedocs.io/en/stable/
django-timezone-field = "^7.0"
# fpdf2 = "^2.8.2" # Probably, will be required later, to generate reports
python = "^3.11"
requests = "^2.32.3"
pyyaml = "^6.0.2"
pillow = "^11.1.0"
django-translated-fields = "^0.13.0"
djangorestframework = "^3.15.2"
# NOTE: The `django-allauth` version somehow resets after adding of a new dependecy, it should be "^65.3.1" (it could be suddenly overriden)
django-allauth = {extras = ["mfa", "socialaccount"], version = "^65.3.1"}
django-cors-headers = "^4.6.0"
django-minify-html = "^1.11.0"
mysqlclient = "^2.2.7"
django-markdownify = "^0.9.5"
django-unfold = "^0.51.0"
django-import-export = "^4.3.7"
django-imagekit = "^5.0.0"
django-ckeditor-5 = "^0.2.17"

[tool.poetry.group.dev.dependencies]
blue = "^0.9.1"
pyright = "^1.1.398"
poetry-plugin-export = "^1.8.0"
isort = "^6.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.autopep8]
max_line_length = 120
# ignore = "E501,W6"  # or ["E501", "W6"]
in-place = true
recursive = true
aggressive = 3

[tool.blue]
line-length = 120
extend-exclude = '''(
\._
| .*\/tales_django\/(migrations)\/.*
)'''

[tool.isort]
# @see https://pycqa.github.io/isort/docs/configuration/options.html
extend_skip = [".venv", "node_modules"]
extend_skip_glob = ["tales_django/migrations/*"]

[tool.poetry.scripts]
export_requirements = "poetry-scripts.export_requirements:export_requirements"
lint = "poetry-scripts:lint"
format = "poetry-scripts:format"
check_all = "poetry-scripts:check_all"
django_migrate = "poetry-scripts.django:migrate"
django_clean_db = "poetry-scripts.django:clean_db"
django_superuser = "poetry-scripts.django:superuser"
django_runserver = "poetry-scripts.django:runserver"
django_livereload = "poetry-scripts.django:livereload"
